services/ztm-hook/requirements.txt
kafka-python==2.0.2
requests==2.25.1
